# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for FECh Plenario
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{id}/attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: Listar asistencia de una sesión
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de asistencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceList"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Marcar asistencia (QR o código)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "200":
          description: Asistencia marcada
        "400":
          description: Código inválido o sesión cerrada
        "409":
          description: Asistencia ya registrada
  /sessions/{id}/attendance/rejoin:
    post:
      operationId: rejoinAttendance
      tags: [attendance]
      summary: Reingresar a la sesión tras haberse retirado
      description: >
        Inverso del auto check-out. Solo funciona mientras la sesión sigue
        abierta: reingresar a una cerrada devolvería el peso de la persona a
        resultados ya fijados.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reingreso registrado
        "400":
          description: La sesión no está abierta, o no estabas retirade
        "404":
          description: Sesión no encontrada
  /sessions/{id}/attendance/checkout:
    post:
      operationId: checkOutAttendance
      tags: [attendance]
      summary: Retirarse de la sesión (auto check-out)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retiro registrado
        "400":
          description: No tienes asistencia activa en esta sesión
  /sessions/{id}/attendance/{userId}:
    patch:
      operationId: updateMemberAttendance
      tags: [attendance]
      summary: Editar asistencia de un miembro (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceEditInput"
      responses:
        "200":
          description: Asistencia actualizada
  /sessions/{id}/attendance/{userId}/checkout:
    post:
      operationId: adminCheckoutMember
      tags: [attendance]
      summary: Retirar a un miembro de la sesión (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Miembro retirado de la sesión
        "400":
          description: El miembro no tiene asistencia activa
  /sessions/{id}/attendance/{userId}/reactivate:
    post:
      operationId: adminReactivateMember
      tags: [attendance]
      summary: Reingresar a un miembro retirado de la sesión (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Miembro reingresado a la sesión
        "400":
          description: El miembro no está retirado de esta sesión
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, username, displayName, group, faculty, votingWeight, votingWeightAlt, active, rol]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        group:
          type: ["string", "null"]
        faculty:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        votingWeight:
          type: number
        votingWeightAlt:
          type: number
        active:
          type: boolean
        rol:
          type: string
...
    AttendanceRecord:
      type: object
      required: [userId, displayName, group, faculty, votingWeight, timestamp, modality]
      properties:
        userId:
          type: integer
        displayName:
          type: string
        group:
          type: ["string", "null"]
        faculty:
          type: ["string", "null"]
        votingWeight:
          type: number
        timestamp:
          type: string
        modality:
          type: string
          enum: [online, presencial]
        checkedOutAt:
          type: ["string", "null"]
    AttendanceList:
      type: object
      required: [present, absent, checkedOut, totalWeight, presentWeight]
      properties:
        present:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        absent:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        checkedOut:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        totalWeight:
          type: number
        presentWeight:
          type: number
        justifiedIds:
          type: array
          description: User ids among `absent` whose absence is marked "Inasistencia Justificada".
          items:
            type: integer
    AttendanceInput:
      type: object
      required: [sessionCode]
      properties:
        sessionCode:
          type: string
        modality:
          type: string
          enum: [online, presencial]
    AttendanceEditInput:
      type: object
      required: [present]
      properties:
        present:
          type: boolean
        modality:
          type: string
          enum: [online, presencial]
        justified:
          type: boolean
          description: Only meaningful with present=false — marks the absence as "Inasistencia Justificada" (label only; still a normal absence for quorum/voting).
...